Refactor: Add an 'admin' package
