+++ sim-inventory/src/main/java/org/ostelco/simcards/admin/SimAdministrationAppConfiguration.kt
@@ -1,4 +1,4 @@
-package org.ostelco.simcards
+package org.ostelco.simcards.admin
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.dropwizard.Configuration
